react-version/backend/requirements.txt
fastapi
uvicorn
pymongo
mercadopago
python-dotenv
yt-dlp
pandas
python-multipart
passlib[bcrypt]
python-jose[cryptography]
google-auth
httpx